% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{final_signature}
\alias{final_signature}
\title{final_signature, the final prognosis-related pathways}
\format{
An object of class \code{character} of length 7.
}
\usage{
final_signature
}
\description{
The final prognosis-related pathways identified by our approach.
}
\keyword{datasets}
