% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_symbol_Entrez}
\alias{gene_symbol_Entrez}
\title{gene_symbol_Entrez, the genes' symbol and ENTREZID}
\format{
An object of class \code{data.frame} with 54245 rows and 2 columns.
}
\usage{
gene_symbol_Entrez
}
\description{
The genes' symbol and ENTREZID.
}
\keyword{datasets}
