% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_Ucox_res}
\alias{gene_Ucox_res}
\title{gene_Ucox_res, the univariate Cox regression result of candidate genes.}
\format{
An object of class \code{data.frame} with 4287 rows and 5 columns.
}
\usage{
gene_Ucox_res
}
\description{
The univariate Cox regression result of candidate genes.
}
\keyword{datasets}
