% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_MultivariateCox_result.R
\name{get_MultivariateCox_result}
\alias{get_MultivariateCox_result}
\title{Perform the multivariate Cox regression}
\usage{
get_MultivariateCox_result(DE_path_sur)
}
\arguments{
\item{DE_path_sur}{A binary metadata table containing sample survival status and survival time.Note that the column names of survival time and survival status must be "survival" and "event".}
}
\value{
Return the multivariate Cox regression results of cancer-specific dysregulated signaling pathways.
}
\description{
The function `get_MultivariateCox_result` uses to perform multivariate Cox regression analysis on the cancer-specific dysregulated signaling pathways.
}
\examples{
#Load the data.
data(path_cox_data)
#perform function `get_MultivariateCox_result`.
res<-get_MultivariateCox_result(path_cox_data)
}
