% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_univarCox_result.R
\name{get_univarCox_result}
\alias{get_univarCox_result}
\title{Perform the univariate Cox regression analysis.}
\usage{
get_univarCox_result(DE_path_sur)
}
\arguments{
\item{DE_path_sur}{A binary metadata table containing  survival status and survival time of each sample.Note that the column names of survival time and survival status must be "survival" and "event"}
}
\value{
Return a data frame, the univariate Cox regression analysis results.
}
\description{
The function `get_univarCox_result` uses to perform the univariate Cox regression analysis.
}
\examples{
#get path of the mutation annotation file.
data(path_cox_data)
#perform function `get_univarCox_result`.
res<-get_univarCox_result(path_cox_data)
}
