% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{km_data}
\alias{km_data}
\title{km_data}
\format{
An object of class \code{data.frame} with 105 rows and 10 columns.
}
\usage{
km_data
}
\description{
The data use for drawing K-M survival curve.
}
\keyword{datasets}
