% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maf_data}
\alias{maf_data}
\title{maf_data}
\format{
An object of class \code{data.frame} with 24461 rows and 4 columns.
}
\usage{
maf_data
}
\description{
The mutation data of patients.
}
\keyword{datasets}
