% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mut_num}
\alias{mut_num}
\title{mut_num}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 13858 rows and 105 columns.
}
\usage{
mut_num
}
\description{
mut_num
}
\keyword{datasets}
