% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mut_sam}
\alias{mut_sam}
\title{mut_sam}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 13858 rows and 2 columns.
}
\usage{
mut_sam
}
\description{
mut_sam.
}
\keyword{datasets}
