% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{path_cox_data}
\alias{path_cox_data}
\title{path_cox_data}
\format{
An object of class \code{data.frame} with 105 rows and 9 columns.
}
\usage{
path_cox_data
}
\description{
path_cox_data
}
\keyword{datasets}
