% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{path_Ucox_mul}
\alias{path_Ucox_mul}
\title{path_Ucox_mul}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 7 rows and 5 columns.
}
\usage{
path_Ucox_mul
}
\description{
path_Ucox_mul
}
\keyword{datasets}
