% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pfs_score}
\alias{pfs_score}
\title{pfs_score}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 123 rows and 105 columns.
}
\usage{
pfs_score
}
\description{
pfs_score.
}
\keyword{datasets}
