% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{roc_data}
\alias{roc_data}
\title{roc_data, the data frame use for ploting ROC curve}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 105 rows and 4 columns.
}
\usage{
roc_data
}
\description{
The roc_data is used to generate ROC curves.
}
\keyword{datasets}
