% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{symbol_Entrez}
\alias{symbol_Entrez}
\title{symbol_Entrez}
\format{
An object of class \code{data.frame} with 54245 rows and 2 columns.
}
\usage{
symbol_Entrez
}
\description{
symbol_Entrez
}
\keyword{datasets}
